Make event merging more intelligent. Encapsulate event dictionaries.
Refactor the event merging system to be more easily modified, and add the functionality of merging velocity changes with position updates.
Change the event constructors to merge the new event into the history immediately, without returning anything. The previous behavior was for the event constructors to return dictionaries that then had to be passed into create_event by the caller.